requirements_cpu.txt
-f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
absl-py>=1.3.0
dm-haiku>=0.0.8
jax>=0.3.24
ml_collections>=0.1.0
numpy>=1.20
typing_extensions>=4.1.1
wrapt>=1.14.1
git+https://github.com/deepmind/einshape@aa4b56d3f850204912511c05de3c4b4e767fb2ce
parameterized>=0.8.1
chex>=0.1.5
